move piotr frontend implementation from game to analyse
