+++ ui/game/src/main.js
@@ -2,7 +2,6 @@ module.exports = {
   game: require('./game'),
   status: require('./status'),
   perf: require('./perf'),
-  piotr: require('./piotr'),
   view: {
     status: require('./view/status'),
     user: require('./view/user'),
